refactor: 상수 데이터 빼기
최대 주문량 상수 데이터 따로 빼기
