+++ src/components/SubmitBtn.jsx
@@ -1,4 +1,4 @@
-import { MAX_ORDER_COUNT } from "../App";
+import { MAX_ORDER_COUNT } from "../constants/order";
 import StyledButton from "./common/StyledButton";
 
 export default function SubmitBtn({ handler, orderCount }) {
