+++ src/constants/order.js
@@ -1,0 +1,1 @@
+export const MAX_ORDER_COUNT = 10;
